added checkboxes to pools filter
