+++ src/components/common/ContentBlock.vue
@@ -25,46 +25,7 @@
           {{ title }}
         </h3>
 
-        <b-dropdown
-          :ref="'dropdown_' + dropdown.id"
-          v-for="dropdown in dropDownFilters"
-          :key="dropdown.id"
-          :text="dropdown.items[dropdown.selectedIndex].title"
-          :variant="
-            !dropDownFiltering(dropdown)
-              ? darkMode
-                ? 'muted-dark'
-                : 'muted-white-light'
-              : darkMode
-              ? 'active-dark'
-              : 'active-light'
-          "
-          :block="true"
-          class="d-none d-lg-inline"
-          toggle-class="block-rounded"
-          :menu-class="
-            darkMode ? 'bg-block-dark shadow' : 'bg-block-light shadow'
-          "
-          style="width: 200px !important; height: 35px !important"
-          :no-caret="true"
-        >
-          <template #button-content>
-            <div class="d-flex justify-content-between align-items-center">
-              {{ dropdown.items[dropdown.selectedIndex].title }}
-              <font-awesome-icon
-                :icon="dropDownFiltering(dropdown) ? 'times' : 'caret-down'"
-                @click.stop="clearFilter(dropdown)"
-              />
-            </div>
-          </template>
-          <b-dropdown-item
-            :class="darkMode ? 'dropdown-item-dark' : 'dropdown-item-light'"
-            v-for="(item, index) in dropdown.items"
-            :key="item.id"
-            @click="dropdown.selectedIndex = index"
-            >{{ item.title }}
-          </b-dropdown-item>
-        </b-dropdown>
+        <slot name="dropDowns"> </slot>
         <div class="d-none d-lg-inline">
           <slot name="date"></slot>
         </div>
@@ -109,10 +70,7 @@
       >
         <slot name="header"> </slot>
       </div>
-      <div
-        v-if="dropDownFilters"
-        class="d-lg-none d-flex align-items-center pb-2 px-3"
-      >
+      <div v-if="filters" class="d-lg-none d-flex align-items-center pb-2 px-3">
         <multi-input-field
           v-if="searchInput !== null"
           class="mr-2"
@@ -122,34 +80,24 @@
           :placeholder="$t('search')"
           prepend="search"
         />
-        <b-btn class="d-flex" variant="primary">
-          <font-awesome-icon icon="filter" @click="showMobileFilters" />
-          <font-awesome-icon
-            icon="times"
-            v-if="anyAreFiltering"
-            class="ml-2"
-            @click="clearAllFilters"
-          />
-        </b-btn>
+        <slot name="mobileFilters"></slot>
       </div>
     </div>
 
     <div class="block-content pb-3 pt-0" :class="px0 ? 'px-0' : ''">
       <slot></slot>
     </div>
-
-    <modal-protected-filters v-model="modal" />
   </div>
 </template>
 <script lang="ts">
 import { Component, Prop, PropSync, Emit } from "vue-property-decorator";
 import MultiInputField from "@/components/common/MultiInputField.vue";
-import ModalProtectedFilters from "@/components/modals/ModalProtectedFilters.vue";
+
 import BaseComponent from "@/components/BaseComponent.vue";
 import { drop } from "node_modules/@types/lodash";
 
 @Component({
-  components: { MultiInputField, ModalProtectedFilters }
+  components: { MultiInputField }
 })
 export default class ContentBlock extends BaseComponent {
   @Prop() title?: string;
@@ -159,20 +107,10 @@ export default class ContentBlock extends BaseComponent {
   @Prop({ default: false }) shadowLight?: boolean;
   @Prop({ default: false }) px0?: boolean;
   @Prop({ default: false }) backButton!: boolean;
-  @Prop() dropDownFilters!: {
-    id: string;
-    selectedIndex: number;
-    items: {
-      title: string;
-    }[];
-  }[];
-  @Prop({ default: false }) dateFiltering!: boolean[];
-  @Prop() clearDate?: Function;
+  @Prop({ default: false }) filters!: boolean;
   @Prop({ default: false }) rippleAnimation?: boolean;
   @PropSync("search", { default: null }) searchInput!: string | null;
   @Prop() searchStyle!: string;
-
-  modal: boolean = false;
 
   @Emit()
   back() {}
@@ -181,34 +119,6 @@ export default class ContentBlock extends BaseComponent {
     const color = this.darkMode ? "text-dark" : "text-light";
     const alignment = this.backButton ? "text-center w-100" : "text-left";
     return [color, alignment];
-  }
-
-  get anyAreFiltering() {
-    let anyFiltering: boolean = false;
-    this.dropDownFilters.map(
-      dropDown => (anyFiltering = dropDown.selectedIndex !== 0)
-    );
-
-    return this.dateFiltering || anyFiltering;
-  }
-
-  showMobileFilters() {
-    this.modal = true;
-  }
-
-  dropDownFiltering(dropDown: any) {
-    return dropDown.selectedIndex !== 0;
-  }
-
-  clearFilter(dropdown: any) {
-    if (this.dropDownFiltering(dropdown)) {
-      dropdown.selectedIndex = 0;
-    }
-  }
-
-  clearAllFilters() {
-    this.dropDownFilters.map(dropDown => dropDown.selectedIndex !== 0);
-    if (this.clearDate) this.clearDate();
   }
 }
 </script>
